Update Ingredient, maxQty field datatype as double from int

+++ app/src/main/java/com/nidhi/recipevault/data/local/model/IngredientModel.kt
@@ -7,7 +7,7 @@ data class IngredientModel(
     @SerializedName("item")
     val item: String,
     @SerializedName("maxQty")
-    val maxQty: Int,
+    val maxQty: Double,
     @SerializedName("qty")
     val qty: Double,
     @SerializedName("unit")
